Static site generator implementation
First let me excuse that this is a huge commit.  I usually try to commit
often and with the least amount of changes.  Anyways, in this case I
developed the site generator while also using it to build another
prototype of my website, so it was in constant change.

With this out of the way let's start with an explanation of the
architecture.  The generator package is the main entry point, its self
named struct provides a Run function that executes three basic steps to
render the actual website:

1. build a library which contains the raw content 2. copy static files
into the output directory 3. render the actual web pages

Point one, building the library, is pretty straightforward and just
walks over all files in the source/content directory.  Note that it is a
bit inconsistenly named, sometimes source and other times content.  For
each file the type is determind, i.e. if its just a plain file or a
page.  Every markdown file is assumed to be a page.  If the file walk
finds a page it will try to parse the front matter and add its metadata
to the library.

The second point is pretty easy and just copies all static files
recursively into the output directory.

Point three, rendering the content, is a bit more tricky.  I started
with a fixed set of templates and assets (CSS files) but later on
decided to allow passing in a different set of templates, e.g. to make
the generator useful for building pages different then my own website.
What the rendering step basically does is to loop over all pages in the
library, converting the markdown content to HTML and rendering the
actual HTML page.  There is some additional logic for sub directories,
if those directories contain pages (markdown files) but no index.md then
a list page is generated, e.g. for a list of all blog articles.

There are two additional packages, frontmatter and slug.  Those packages
were copied from a previous iteration of my website server, which is
unreleased.  Package frontmatter contains logic for parsing page front
matters.  Despite the commonly used delimiters it simply expects that
the file begins with a markdown code block where the contents are
interpreted as the contents of the front matter, e.g. page title, author
tags etc.  This has the advantage that it does not break syntax
highlighting in contrast to other formats and it was easy to parse.  The
slug package provides a slugifier which is just a function that can
convert arbitrary unicode page titles into slugs (fragments) that can
safely be used in a URL.  Note that the file name of a rendered HTML
page is the slugified title of its front matter.