Parallelize file copying
This parallelizes copying of static and other assets.
The level of concurrency is not parameterized yet and defaults to the
number of CPU cores available.

Running the static assets copy benchmark gives the following results:

    $ go test -bench=BenchmarkCopyStaticFiles ./generator/
    goos: linux
    goarch: amd64
    pkg: github.com/klingtnet/static-site-generator/generator
    cpu: AMD Ryzen 7 3700X 8-Core Processor
    BenchmarkCopyStaticFiles-16                   1        1116907777 ns/op
    PASS
    ok      github.com/klingtnet/static-site-generator/generator    1.155s

There almost no visible speed up mainly because those are just in memory
copy operationbs due to the testfs.MapFS and the DiscardStorage.  Adding
a short delay to the DiscardStorage.Store method, e.g. 10ms, shows a
different result and then the parallelized version is much faster, about
10x in the case of a 10ms delay.