Setup GitHub Pages
This will generate a single file site using REAMDE.md as the root index file.

The setup-gh-pages branch will be removed from the list of gh-pages
branches after this gets merged.

Note that the source branch for the website must be configured in:
https://github.com/klingtnet/static-site-generator/settings/pages

Documentation about GitHub Pages can be found here:
https://docs.github.com/en/pages/getting-started-with-github-pages/about-github-pages