Use a tree data structure for content library
Storing the content in a tree like data structure feels more natural
since the content is already stored in a hierarchical directory
structure.  This also simplifies the implementation in the sense that
building the library only requires to traverse the tree.  Still, there
is some special handling since I only want to support a root website
menu, but if we need to, we could easily generate sub menus and the like
without changing much of the logic.

This commit also moved the website data model into a separate sub
package which helps to clarify the code structure further.  A similar
change was applied to the renderer by moving it into the renderer
package and by defining a Renderer interface to allow for easy mocking
inside a test.

Note that the generator got a tiny bit slower than before:

    $ go test -benchmem -run=^$ -bench ^BenchmarkGenerator$ -v github.com/klingtnet/static-site-generator/generator
    goos: linux
    goarch: amd64
    pkg: github.com/klingtnet/static-site-generator/generator
    cpu: AMD Ryzen 7 3700X 8-Core Processor
    BenchmarkGenerator
    BenchmarkGenerator/concurrency-1
    BenchmarkGenerator/concurrency-1-16                    1        1387738786 ns/op        204529032 B/op   5031689 allocs/op
    BenchmarkGenerator/concurrency-16
    BenchmarkGenerator/concurrency-16-16                   4         269390143 ns/op        210134488 B/op   5032763 allocs/op
    BenchmarkGenerator/concurrency-32
    BenchmarkGenerator/concurrency-32-16                   4         265208211 ns/op        212981888 B/op   5033533 allocs/op
    PASS
    ok      github.com/klingtnet/static-site-generator/generator    5.753s

Benchmark results cannot be compared one to one because the benchmark was
changed such that the content is now deeply nested.  This is something
the old data structure did not support very well.