Add livereloading
This adds a new livereload subcommand that will rebuild the website
anytime a resource has changed.  Only fs.FS are used internally by the
generator, hence I could not rely on something like fsnotify to watch
for changes.  Instead, I needed to check the resource fs's in regular
intervals.  This is not very efficient but was straightforward to
implement and is still more than efficient enough, depending on the
size of the input resources.  Also, the watcher will work for any kind
of fs.FS, e.g. for an s3 of SFTP abstraction.