Render RSS feeds
This adds RSS feed rendering for all directories that contain articles,
except the root directory.  Feeds are always stores in <dir>/feed.rss.
We now need to pass the config into Generator since we need the base URL
and author information to provide proper feed metadata.  A new template
is required as well, feed.gohtml, which replaces base.gohtml for feed
rendering and should just omit header and footer parts of the page since
we don't need a menu and footer information for those feed pages.

I saw a bunch of context.TODO that should be replaced by passing down
the root context.