Add and use concurrency pattern package
This adds a new internal package called distribute.  At the moment only
a single function called OneToN is defined in this package.  It
implements the concurrency pattern that was used in many places in the
site generator, that is distribution of work to a limited amount of
workers (limited concurrency).  Also, normal pages and list pages as
well as feeds are now rendered in parallel.

	goos: darwin
	goarch: arm64
	pkg: github.com/klingtnet/static-site-generator/generator
	BenchmarkGenerator/concurrency-1-8         	       3	 498528569 ns/op
	BenchmarkGenerator/concurrency-4-8         	       6	 179463861 ns/op
	BenchmarkGenerator/concurrency-8-8         	       6	 175925438 ns/op
	BenchmarkGenerator/concurrency-16-8        	       6	 174429535 ns/op

This is a tiny bit faster than before but still slower than without
rendering feeds.  When feed rendering is disabled (by removing the call)
we get the following results:

	goos: darwin
	goarch: arm64
	pkg: github.com/klingtnet/static-site-generator/generator
	BenchmarkGenerator/concurrency-1-8         	       7	 146820059 ns/op
	BenchmarkGenerator/concurrency-4-8         	      21	  56286597 ns/op
	BenchmarkGenerator/concurrency-8-8         	      21	  54206764 ns/op
	BenchmarkGenerator/concurrency-16-8        	      21	  55374026 ns/op

About one third faster than without the use of OneToN, which is good.
Still, it's fast enough with feed rendering enabled.  I thought about
what optimizations might be feasible and could rule out caching since we
do not render anything twice (this might need to be verified).

Closes #5.

This also fixes two small issues, one was a leftover TODO context and
the other was a nil initialized buffer for feed page rendering that is
now set up with a size of 8kB.